+++ ui/WLoadedEsquemes.h
@@ -1,7 +1,8 @@
-// =================================================== \\
-// ====     Copyright (c) 2024 Ignasi Camps       ==== \\
-// ==== SPDX-License-Identifier: GPL-3.0-or-later ==== \\
-// =================================================== \\
+/* =================================================== *
+ * ====        Copyright (c) 2024 icampsi         ==== *
+
+ * ==== SPDX-License-Identifier: GPL-3.0-or-later ==== *
+ * =================================================== */
 
 #ifndef WLOADEDESQUEMES_H
 #define WLOADEDESQUEMES_H
@@ -20,7 +21,7 @@ public:
     explicit WLoadedEsquemes(QWidget *parent = nullptr);
     ~WLoadedEsquemes();
     void newEsquema(PEsquemaPage *page, CEsquema *esquema);
-    void deleteEsquema(const int index) { ui->list_esquemes->handleDeleteEsquema(index); }
+    void deleteEsquema(const size_t index) { ui->list_esquemes->handleDeleteEsquema(index); }
 
 private slots:
     void on_list_esquemes_itemSelectionChanged();
